javascript cleanup
